+++ __tests__/components/Button.test.tsx
@@ -1,4 +1,4 @@
-import Button from '@/components/buttons';
+import Button from '@/components/Button';
 import { fireEvent, render, screen } from '@testing-library/react';
 
 jest.mock('@/libs/utils/Icon', () => ({
