+++ src/components/commons/PageLayouts/index.tsx
@@ -3,8 +3,8 @@ import { usePathname } from 'next/navigation';
 import { useRouter } from 'next/router';
 import { ReactNode } from 'react';
 
+import Footer from '../Footer';
 import Header from '../Header';
-import Footer from '../footers';
 
 type LayoutProps = {
   children: ReactNode;
