+++ src/pages/_app.tsx
@@ -1,4 +1,4 @@
-import Layout from '@/components/PageLayouts/pageLayout';
+import Layout from '@/components/commons/PageLayouts';
 import '@/styles/globals.css';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
